src/main/java/hello/mvc2/domain/item/ItemRepository.java
package hello.mvc2.domain.item;

import hello.mvc2.domain.item.Item;
import org.springframework.stereotype.Repository;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

@Repository
public class ItemRepository {

    private static Map<Long, Item> store = new HashMap<>();
    private static long sequence=0L;

    public Item save(Item item){
        item.setId(++sequence);
        store.put(item.getId(),item);
        return item;
    }

    public Item findById(Long id){
        return store.get(id);
    }

    public List<Item> findAll(){
        return new ArrayList<>(store.values());
    }

    public void update(Long itemId, Item updateParam){
        Item item = findById(itemId);
        item.setItemName(updateParam.getItemName());
        item.setPrice(updateParam.getPrice());
        item.setQuantity(updateParam.getQuantity());
        item.setOpen(updateParam.getOpen());
        item.setRegions(updateParam.getRegions());
        item.setItemType(updateParam.getItemType());
        item.setDeliveryCode(updateParam.getDeliveryCode());
    }

    public void clearStore(){
        store.clear();
    }

}
